FEATURE: add internal code
This is an initial commit that adds all code from the internal extension
removing all commit history as well as dependencies to other internal
extensions.

This is a new major release starting with version 4.x to prevent
conflicts with the old internal extension.

This will be the basis for further development.